Refactor ComponentAnalysisPlotExportDialog to a tabbed pane in ComponentAnalysisDialog
