+++ core/src/main/resources/l10n/messages.properties
@@ -909,44 +909,48 @@ compaddbuttons.askPosition.Inserthere = Insert here
 compaddbuttons.askPosition.Addtotheend = Add to the end
 compaddbuttons.askPosition.Cancel = Cancel
 
-! Component Analysis Dialog
-componentanalysisdlg.componentanalysis = Component analysis
-componentanalysisdlg.lbl.winddir = Wind direction:
-componentanalysisdlg.TitledBorder.warnings = Warnings:
-componentanalysisdlg.ToggleBut.worst = Worst
-componentanalysisdlg.lbl.angleofattack = Angle of attack:
-componentanalysisdlg.lbl.machnumber = Mach number:
-componentanalysisdlg.lbl.rollrate = Roll rate:
-componentanalysisdlg.lbl.activestages = Active stages:
-componentanalysisdlg.lbl.motorconf = Motor configuration:
-componentanalysisdlg.TabStability.Col = Component Name
-componentanalysisdlg.TabStability.Col.EachMass = Instance Mass
-componentanalysisdlg.TabStability.Col.AllMass = Aggregate Mass
-componentanalysisdlg.TabStability.Col.CG = Aggregate CG Location
-componentanalysisdlg.TabStability.Col.CP = CP
-componentanalysisdlg.TabStability = Stability
-componentanalysisdlg.TabStability.ttip = Stability information
-componentanalysisdlg.dragTableModel.Col.Component = Component
-componentanalysisdlg.dragTableModel.Col.Pressure = <html>Pressure C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Base = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction = <html>Friction C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total = <html>Total C<sub>D</sub>
-componentanalysisdlg.dragTabchar = Drag characteristics
-componentanalysisdlg.dragTabchar.ttip = Drag characteristics
-componentanalysisdlg.rollTableModel.Col.component = Component
-componentanalysisdlg.rollTableModel.Col.rollforc = Roll forcing coefficient
-componentanalysisdlg.rollTableModel.Col.rolldamp = Roll damping coefficient
-componentanalysisdlg.rollTableModel.Col.total = <html>Total C<sub>l</sub>
-componentanalysisdlg.rollTableModel = Roll dynamics
-componentanalysisdlg.rollTableModel.ttip = Roll dynamics
-componentanalysisdlg.println.closingmethod = Closing method called:
-componentanalysisdlg.println.settingnam = SETTING NAN VALUES
-componentanalysisdlg.lbl.reflenght = Reference length:
-componentanalysisdlg.lbl.refarea = Reference area:
+! ComponentAnalysisDialog
+ComponentAnalysisDialog.componentanalysis = Component analysis
+ComponentAnalysisDialog.tab.General = General
+ComponentAnalysisDialog.tab.PlotExport = Plot / Export
+
+! ComponentAnalysisGeneralTab
+ComponentAnalysisGeneralTab.lbl.winddir = Wind direction:
+ComponentAnalysisGeneralTab.TitledBorder.warnings = Warnings:
+ComponentAnalysisGeneralTab.ToggleBut.worst = Worst
+ComponentAnalysisGeneralTab.lbl.angleofattack = Angle of attack:
+ComponentAnalysisGeneralTab.lbl.machnumber = Mach number:
+ComponentAnalysisGeneralTab.lbl.rollrate = Roll rate:
+ComponentAnalysisGeneralTab.lbl.activestages = Active stages:
+ComponentAnalysisGeneralTab.lbl.motorconf = Motor configuration:
+ComponentAnalysisGeneralTab.TabStability.Col = Component Name
+ComponentAnalysisGeneralTab.TabStability.Col.EachMass = Instance Mass
+ComponentAnalysisGeneralTab.TabStability.Col.AllMass = Aggregate Mass
+ComponentAnalysisGeneralTab.TabStability.Col.CG = Aggregate CG Location
+ComponentAnalysisGeneralTab.TabStability.Col.CP = CP
+ComponentAnalysisGeneralTab.TabStability = Stability
+ComponentAnalysisGeneralTab.TabStability.ttip = Stability information
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Component
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure = <html>Pressure C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction = <html>Friction C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total = <html>Total C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTabchar = Drag characteristics
+ComponentAnalysisGeneralTab.dragTabchar.ttip = Drag characteristics
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Component
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc = Roll forcing coefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp = Roll damping coefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.total = <html>Total C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel = Roll dynamics
+ComponentAnalysisGeneralTab.rollTableModel.ttip = Roll dynamics
+ComponentAnalysisGeneralTab.println.closingmethod = Closing method called:
+ComponentAnalysisGeneralTab.println.settingnam = SETTING NAN VALUES
+ComponentAnalysisGeneralTab.lbl.reflenght = Reference length:
+ComponentAnalysisGeneralTab.lbl.refarea = Reference area:
 !componentanalysisdlg.But.close =Close
-componentanalysisdlg.TabStability.Col.Component = Component
-componentanalysisdlg.TOTAL = Total (Rocket)
-componentanalysisdlg.noWarnings = <html><i><font color=\"gray\">No warnings.</font></i>
+ComponentAnalysisGeneralTab.TabStability.Col.Component = Component
+ComponentAnalysisGeneralTab.TOTAL = Total (Rocket)
+ComponentAnalysisGeneralTab.noWarnings = <html><i><font color=\"gray\">No warnings.</font></i>
 
 ! CAPlotExportDialog
 CAPlotExportDialog.title = Plot / Export Component Analysis
