+++ core/src/main/resources/l10n/messages_ar.properties
@@ -747,43 +747,43 @@ compaddbuttons.askPosition.Addtotheend = إضف للنهاية
 compaddbuttons.askPosition.Cancel = إلغاء
 
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis = تحليل المكونات
-componentanalysisdlg.lbl.winddir = إتجاه الريح
-componentanalysisdlg.TitledBorder.warnings = :تحذيرات
-componentanalysisdlg.ToggleBut.worst = أسوأ
-componentanalysisdlg.lbl.angleofattack = :زاوية المواجهة
-componentanalysisdlg.lbl.machnumber = :رقم ماخ
-componentanalysisdlg.lbl.rollrate = :معدل الإلتفاف
-componentanalysisdlg.lbl.activestages = :المراحل النشطة
-componentanalysisdlg.lbl.motorconf = :إعدادات المحرك
-componentanalysisdlg.TabStability.Col = إسم المكون
-componentanalysisdlg.TabStability.Col.EachMass = كتلة المثيل
-componentanalysisdlg.TabStability.Col.AllMass = الكتلة الكلية
-componentanalysisdlg.TabStability.Col.CG = موقع مركز الجاذبية الكلي
-componentanalysisdlg.TabStability.Col.CP = مركز الضغط
-componentanalysisdlg.TabStability = الإستقرار
-componentanalysisdlg.TabStability.ttip = معلومات الإستقرار
-componentanalysisdlg.dragTableModel.Col.Component = مكون
-componentanalysisdlg.dragTableModel.Col.Pressure = <html>مركز الضغط</sub>
-componentanalysisdlg.dragTableModel.Col.Base = <html>معامل السحب الأساسي</sub>
-componentanalysisdlg.dragTableModel.Col.friction = <html>معامل مقاومة الإحتكاك</sub>
-componentanalysisdlg.dragTableModel.Col.total = <html>معامل السحب الكلي</sub>
-componentanalysisdlg.dragTabchar = خصائص السحب
-componentanalysisdlg.dragTabchar.ttip = خصائص السحب
-componentanalysisdlg.rollTableModel.Col.component = المكون
-componentanalysisdlg.rollTableModel.Col.rollforc = معامل تأثير الإلتفاف
-componentanalysisdlg.rollTableModel.Col.rolldamp = معامل تخميد الإلتفاف
-componentanalysisdlg.rollTableModel.Col.total = <html>إجمالي معامل الإلتفاف
-componentanalysisdlg.rollTableModel = ديناميات الإلتفاف
-componentanalysisdlg.rollTableModel.ttip = ديناميات الإلتفاف
-componentanalysisdlg.println.closingmethod = :طريقة الإغلاق تسمى
-componentanalysisdlg.println.settingnam = عدم ضبط قيم الأرقام
-componentanalysisdlg.lbl.reflenght = :طول مرجعي
-componentanalysisdlg.lbl.refarea = :المنطقة المرجعية
+ComponentAnalysisDialog.componentanalysis = تحليل المكونات
+ComponentAnalysisGeneralTab.lbl.winddir = إتجاه الريح
+ComponentAnalysisGeneralTab.TitledBorder.warnings = :تحذيرات
+ComponentAnalysisGeneralTab.ToggleBut.worst = أسوأ
+ComponentAnalysisGeneralTab.lbl.angleofattack = :زاوية المواجهة
+ComponentAnalysisGeneralTab.lbl.machnumber = :رقم ماخ
+ComponentAnalysisGeneralTab.lbl.rollrate = :معدل الإلتفاف
+ComponentAnalysisGeneralTab.lbl.activestages = :المراحل النشطة
+ComponentAnalysisGeneralTab.lbl.motorconf = :إعدادات المحرك
+ComponentAnalysisGeneralTab.TabStability.Col = إسم المكون
+ComponentAnalysisGeneralTab.TabStability.Col.EachMass = كتلة المثيل
+ComponentAnalysisGeneralTab.TabStability.Col.AllMass = الكتلة الكلية
+ComponentAnalysisGeneralTab.TabStability.Col.CG = موقع مركز الجاذبية الكلي
+ComponentAnalysisGeneralTab.TabStability.Col.CP = مركز الضغط
+ComponentAnalysisGeneralTab.TabStability = الإستقرار
+ComponentAnalysisGeneralTab.TabStability.ttip = معلومات الإستقرار
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = مكون
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure = <html>مركز الضغط</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base = <html>معامل السحب الأساسي</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction = <html>معامل مقاومة الإحتكاك</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total = <html>معامل السحب الكلي</sub>
+ComponentAnalysisGeneralTab.dragTabchar = خصائص السحب
+ComponentAnalysisGeneralTab.dragTabchar.ttip = خصائص السحب
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = المكون
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc = معامل تأثير الإلتفاف
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp = معامل تخميد الإلتفاف
+ComponentAnalysisGeneralTab.rollTableModel.Col.total = <html>إجمالي معامل الإلتفاف
+ComponentAnalysisGeneralTab.rollTableModel = ديناميات الإلتفاف
+ComponentAnalysisGeneralTab.rollTableModel.ttip = ديناميات الإلتفاف
+ComponentAnalysisGeneralTab.println.closingmethod = :طريقة الإغلاق تسمى
+ComponentAnalysisGeneralTab.println.settingnam = عدم ضبط قيم الأرقام
+ComponentAnalysisGeneralTab.lbl.reflenght = :طول مرجعي
+ComponentAnalysisGeneralTab.lbl.refarea = :المنطقة المرجعية
 !componentanalysisdlg.But.close = إغلاق
-componentanalysisdlg.TabStability.Col.Component = المكون
-componentanalysisdlg.TOTAL = الكلي
-componentanalysisdlg.noWarnings = <html><i><font color=\"gray\">لا تحذيرات</font></i>
+ComponentAnalysisGeneralTab.TabStability.Col.Component = المكون
+ComponentAnalysisGeneralTab.TOTAL = الكلي
+ComponentAnalysisGeneralTab.noWarnings = <html><i><font color=\"gray\">لا تحذيرات</font></i>
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir = إتجاه الريح
