+++ core/src/main/resources/l10n/messages_cs.properties
@@ -508,37 +508,37 @@ compaddbuttons.askPosition.Addtotheend = Pridej na konec
 compaddbuttons.askPosition.Cancel = Zru\u0161it
 
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis = Analýza soucástky
-componentanalysisdlg.lbl.winddir = Smer vetru:
-componentanalysisdlg.TitledBorder.warnings = Varování:
-componentanalysisdlg.ToggleBut.worst = Nejhor\u0161í
-componentanalysisdlg.lbl.angleofattack = Úhel nábehu:
-componentanalysisdlg.lbl.machnumber = Machovo císlo:
-componentanalysisdlg.lbl.rollrate = Rychlost otácení:
-componentanalysisdlg.lbl.activestages = Startovací rampa:
-componentanalysisdlg.lbl.motorconf = Nastavení motoru:
-componentanalysisdlg.TabStability.Col = Komponenta
-componentanalysisdlg.TabStability = Stabilita
-componentanalysisdlg.TabStability.ttip = Informace o stabilite
-componentanalysisdlg.dragTableModel.Col.Component = Komponenta
-componentanalysisdlg.dragTableModel.Col.Pressure = <html>Tlak C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Base = <html>Základní C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction = <html>Trení C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total = <html>Celkem C<sub>D</sub>
-componentanalysisdlg.dragTabchar = Vlastnosti trupu
-componentanalysisdlg.dragTabchar.ttip = Drag characteristics
-componentanalysisdlg.rollTableModel.Col.component = Komponenta
-componentanalysisdlg.rollTableModel.Col.rollforc = Otácecí koefficient
-componentanalysisdlg.rollTableModel.Col.rolldamp = Otácecí tlumící coefficient
-componentanalysisdlg.rollTableModel.Col.total = <html>Celkem C<sub>l</sub>
-componentanalysisdlg.rollTableModel = Otácecí dynamika
-componentanalysisdlg.rollTableModel.ttip = Otácecí dynamika
-componentanalysisdlg.println.closingmethod = Jméno uzavírací metody:
-componentanalysisdlg.println.settingnam = Nastavení NAN hodnot
-componentanalysisdlg.lbl.reflenght = Doporucená délka: 
-componentanalysisdlg.lbl.refarea = Doporucená plocha: 
+ComponentAnalysisDialog.componentanalysis = Analýza soucástky
+ComponentAnalysisGeneralTab.lbl.winddir = Smer vetru:
+ComponentAnalysisGeneralTab.TitledBorder.warnings = Varování:
+ComponentAnalysisGeneralTab.ToggleBut.worst = Nejhor\u0161í
+ComponentAnalysisGeneralTab.lbl.angleofattack = Úhel nábehu:
+ComponentAnalysisGeneralTab.lbl.machnumber = Machovo císlo:
+ComponentAnalysisGeneralTab.lbl.rollrate = Rychlost otácení:
+ComponentAnalysisGeneralTab.lbl.activestages = Startovací rampa:
+ComponentAnalysisGeneralTab.lbl.motorconf = Nastavení motoru:
+ComponentAnalysisGeneralTab.TabStability.Col = Komponenta
+ComponentAnalysisGeneralTab.TabStability = Stabilita
+ComponentAnalysisGeneralTab.TabStability.ttip = Informace o stabilite
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Komponenta
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure = <html>Tlak C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base = <html>Základní C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction = <html>Trení C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total = <html>Celkem C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTabchar = Vlastnosti trupu
+ComponentAnalysisGeneralTab.dragTabchar.ttip = Drag characteristics
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Komponenta
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc = Otácecí koefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp = Otácecí tlumící coefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.total = <html>Celkem C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel = Otácecí dynamika
+ComponentAnalysisGeneralTab.rollTableModel.ttip = Otácecí dynamika
+ComponentAnalysisGeneralTab.println.closingmethod = Jméno uzavírací metody:
+ComponentAnalysisGeneralTab.println.settingnam = Nastavení NAN hodnot
+ComponentAnalysisGeneralTab.lbl.reflenght = Doporucená délka: 
+ComponentAnalysisGeneralTab.lbl.refarea = Doporucená plocha: 
 !componentanalysisdlg.But.close = Zavrít
-componentanalysisdlg.TabStability.Col.Component = Komponenta
+ComponentAnalysisGeneralTab.TabStability.Col.Component = Komponenta
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir = Smer vetru
