+++ core/src/main/resources/l10n/messages_de.properties
@@ -565,37 +565,37 @@ compaddbuttons.askPosition.Addtotheend = Zum Ende hinzufügen
 compaddbuttons.askPosition.Cancel = Abbrechen
 
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis = Component analysis
-componentanalysisdlg.lbl.winddir = Windrichtung
-componentanalysisdlg.TitledBorder.warnings = Warnungen:
-componentanalysisdlg.ToggleBut.worst = Worst
-componentanalysisdlg.lbl.angleofattack = Angriffswinkel:
-componentanalysisdlg.lbl.machnumber = Machzahl:
-componentanalysisdlg.lbl.rollrate = Rollrate:
-componentanalysisdlg.lbl.activestages = Aktive Stufen:
-componentanalysisdlg.lbl.motorconf = Motorkonfiguration:
-componentanalysisdlg.TabStability.Col = Komponente
-componentanalysisdlg.TabStability = Stabilität
-componentanalysisdlg.TabStability.ttip = Stabilitätsinformationen
-componentanalysisdlg.dragTableModel.Col.Component = Komponente
-componentanalysisdlg.dragTableModel.Col.Pressure = <html>Druck C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Base = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction = <html>Reibung C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total = <html>Gesamt C<sub>D</sub>
-componentanalysisdlg.dragTabchar = Luftwiderstand
-componentanalysisdlg.dragTabchar.ttip = Luftwiderstands-Eigenschaften
-componentanalysisdlg.rollTableModel.Col.component = Komponente
-componentanalysisdlg.rollTableModel.Col.rollforc = Roll forcing coefficient
-componentanalysisdlg.rollTableModel.Col.rolldamp = Roll damping coefficient
-componentanalysisdlg.rollTableModel.Col.total = <html>Gesamt C<sub>l</sub>
-componentanalysisdlg.rollTableModel = Rolldynamik
-componentanalysisdlg.rollTableModel.ttip = Rolldynamik
-componentanalysisdlg.println.closingmethod = Closing method called:
-componentanalysisdlg.println.settingnam = SETTING NAN VALUES
-componentanalysisdlg.lbl.reflenght = Referenzlänge:
-componentanalysisdlg.lbl.refarea = Referenzfläche:
+ComponentAnalysisDialog.componentanalysis = Component analysis
+ComponentAnalysisGeneralTab.lbl.winddir = Windrichtung
+ComponentAnalysisGeneralTab.TitledBorder.warnings = Warnungen:
+ComponentAnalysisGeneralTab.ToggleBut.worst = Worst
+ComponentAnalysisGeneralTab.lbl.angleofattack = Angriffswinkel:
+ComponentAnalysisGeneralTab.lbl.machnumber = Machzahl:
+ComponentAnalysisGeneralTab.lbl.rollrate = Rollrate:
+ComponentAnalysisGeneralTab.lbl.activestages = Aktive Stufen:
+ComponentAnalysisGeneralTab.lbl.motorconf = Motorkonfiguration:
+ComponentAnalysisGeneralTab.TabStability.Col = Komponente
+ComponentAnalysisGeneralTab.TabStability = Stabilität
+ComponentAnalysisGeneralTab.TabStability.ttip = Stabilitätsinformationen
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Komponente
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure = <html>Druck C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction = <html>Reibung C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total = <html>Gesamt C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTabchar = Luftwiderstand
+ComponentAnalysisGeneralTab.dragTabchar.ttip = Luftwiderstands-Eigenschaften
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Komponente
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc = Roll forcing coefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp = Roll damping coefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.total = <html>Gesamt C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel = Rolldynamik
+ComponentAnalysisGeneralTab.rollTableModel.ttip = Rolldynamik
+ComponentAnalysisGeneralTab.println.closingmethod = Closing method called:
+ComponentAnalysisGeneralTab.println.settingnam = SETTING NAN VALUES
+ComponentAnalysisGeneralTab.lbl.reflenght = Referenzlänge:
+ComponentAnalysisGeneralTab.lbl.refarea = Referenzfläche:
 !componentanalysisdlg.But.close =Close
-componentanalysisdlg.TabStability.Col.Component = Komponente
+ComponentAnalysisGeneralTab.TabStability.Col.Component = Komponente
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir = Windrichtung
