+++ core/src/main/resources/l10n/messages_es.properties
@@ -1282,43 +1282,43 @@ compaddbuttons.askPosition.Inserthere  = Insertar aqu\u00ed
 compaddbuttons.lbl.Youcanchange        = Puede cambiar la operaci\u00f3n con las preferencias por defecto
 compaddbuttons.lbl.insertcomp          = \u00bfInsertar el componente despu\u00e9s del actual o al final?
 
-componentanalysisdlg.TOTAL                        = Total
-componentanalysisdlg.TabStability                 = Estabilidad
-componentanalysisdlg.TabStability.Col             = Componente
-componentanalysisdlg.TabStability.Col.CG          = CG
-componentanalysisdlg.TabStability.Col.CP          = CP
+ComponentAnalysisGeneralTab.TOTAL                        = Total
+ComponentAnalysisGeneralTab.TabStability                 = Estabilidad
+ComponentAnalysisGeneralTab.TabStability.Col             = Componente
+ComponentAnalysisGeneralTab.TabStability.Col.CG          = CG
+ComponentAnalysisGeneralTab.TabStability.Col.CP          = CP
 !componentanalysisdlg.But.close = Cerrar
-componentanalysisdlg.TabStability.Col.Component   = Componente
+ComponentAnalysisGeneralTab.TabStability.Col.Component   = Componente
 componentanalysisdlg.TabStability.Col.Mass        = Masa
-componentanalysisdlg.TabStability.ttip            = Informaci\u00f3n de Estabilidad
-componentanalysisdlg.TitledBorder.warnings        = Advertencias:
-componentanalysisdlg.ToggleBut.worst              = Peor
+ComponentAnalysisGeneralTab.TabStability.ttip            = Informaci\u00f3n de Estabilidad
+ComponentAnalysisGeneralTab.TitledBorder.warnings        = Advertencias:
+ComponentAnalysisGeneralTab.ToggleBut.worst              = Peor
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis            = An\u00e1lisis de los componentes
-componentanalysisdlg.dragTabchar                  = Caracter\u00edsticas de rozamiento
-componentanalysisdlg.dragTabchar.ttip             = Coeficientes de arrastre de los componentes.
-componentanalysisdlg.dragTableModel.Col.Base      = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Component = Componente
-componentanalysisdlg.dragTableModel.Col.Pressure  = <html>Presi\u00f3n C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction  = <html>Rozamiento C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total     = <html>Total C<sub>D</sub>
-componentanalysisdlg.lbl.activestages             = Etapas activas:
-componentanalysisdlg.lbl.angleofattack            = \u00c1ngulo de ataque:
-componentanalysisdlg.lbl.machnumber               = N\u00famero Mach:
-componentanalysisdlg.lbl.motorconf                = Configuraci\u00f3n del motor:
-componentanalysisdlg.lbl.refarea                  = \u00c1rea de referencia: 
-componentanalysisdlg.lbl.reflenght                = Di\u00e1metro de referencia: 
-componentanalysisdlg.lbl.rollrate                 = Valor de rotaci\u00f3n:
-componentanalysisdlg.lbl.winddir                  = Direcci\u00f3n del viento:
-componentanalysisdlg.noWarnings                   = <html><i><font color="gray">Sin advertencias.</font></i>
-componentanalysisdlg.println.closingmethod        = Llamar al m\u00e9todo de cierre:
-componentanalysisdlg.println.settingnam           = AJUSTANDO VALORES NAN
-componentanalysisdlg.rollTableModel               = Din\u00e1mica de rotaci\u00f3n
-componentanalysisdlg.rollTableModel.Col.component = Componente
-componentanalysisdlg.rollTableModel.Col.rolldamp  = Coeficiente de correcci\u00f3n
-componentanalysisdlg.rollTableModel.Col.rollforc  = Coeficiente de rotaci\u00f3n
-componentanalysisdlg.rollTableModel.Col.total     = <html>Total C<sub>l</sub>
-componentanalysisdlg.rollTableModel.ttip          = Din\u00e1mica del movimiento de rotaci\u00f3n del cohete (spin)
+ComponentAnalysisDialog.componentanalysis            = An\u00e1lisis de los componentes
+ComponentAnalysisGeneralTab.dragTabchar                  = Caracter\u00edsticas de rozamiento
+ComponentAnalysisGeneralTab.dragTabchar.ttip             = Coeficientes de arrastre de los componentes.
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base      = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Componente
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure  = <html>Presi\u00f3n C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction  = <html>Rozamiento C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total     = <html>Total C<sub>D</sub>
+ComponentAnalysisGeneralTab.lbl.activestages             = Etapas activas:
+ComponentAnalysisGeneralTab.lbl.angleofattack            = \u00c1ngulo de ataque:
+ComponentAnalysisGeneralTab.lbl.machnumber               = N\u00famero Mach:
+ComponentAnalysisGeneralTab.lbl.motorconf                = Configuraci\u00f3n del motor:
+ComponentAnalysisGeneralTab.lbl.refarea                  = \u00c1rea de referencia: 
+ComponentAnalysisGeneralTab.lbl.reflenght                = Di\u00e1metro de referencia: 
+ComponentAnalysisGeneralTab.lbl.rollrate                 = Valor de rotaci\u00f3n:
+ComponentAnalysisGeneralTab.lbl.winddir                  = Direcci\u00f3n del viento:
+ComponentAnalysisGeneralTab.noWarnings                   = <html><i><font color="gray">Sin advertencias.</font></i>
+ComponentAnalysisGeneralTab.println.closingmethod        = Llamar al m\u00e9todo de cierre:
+ComponentAnalysisGeneralTab.println.settingnam           = AJUSTANDO VALORES NAN
+ComponentAnalysisGeneralTab.rollTableModel               = Din\u00e1mica de rotaci\u00f3n
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Componente
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp  = Coeficiente de correcci\u00f3n
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc  = Coeficiente de rotaci\u00f3n
+ComponentAnalysisGeneralTab.rollTableModel.Col.total     = <html>Total C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel.ttip          = Din\u00e1mica del movimiento de rotaci\u00f3n del cohete (spin)
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir                  = Direcci\u00f3n del viento
