+++ core/src/main/resources/l10n/messages_fr.properties
@@ -1276,44 +1276,44 @@ compaddbuttons.askPosition.Inserthere  = Ajouter ici
 compaddbuttons.lbl.Youcanchange        = Vous pouvez modifier le fonctionnement par d\u00E9faut dans les pr\u00E9f\u00E9rences.
 compaddbuttons.lbl.insertcomp          = Ins\u00E9rez la pi\u00E8ce apr\u00E8s l'\u00E9l\u00E9ment actuel ou en tant que dernier \u00E9l\u00E9ment?
 
-componentanalysisdlg.TOTAL                        = Total
-componentanalysisdlg.TabStability                 = Stabilit\u00E9
+ComponentAnalysisGeneralTab.TOTAL                        = Total
+ComponentAnalysisGeneralTab.TabStability                 = Stabilit\u00E9
 !componentanalysisdlg.TabStability.Col.Component = Pi\u00e8ce
-componentanalysisdlg.TabStability.Col             = Pi\u00E8ce
-componentanalysisdlg.TabStability.Col.CG          = CG
-componentanalysisdlg.TabStability.Col.CP          = CP
+ComponentAnalysisGeneralTab.TabStability.Col             = Pi\u00E8ce
+ComponentAnalysisGeneralTab.TabStability.Col.CG          = CG
+ComponentAnalysisGeneralTab.TabStability.Col.CP          = CP
 !componentanalysisdlg.but.close =Fermer
-componentanalysisdlg.TabStability.Col.Component   = Pi\u00E8ce
+ComponentAnalysisGeneralTab.TabStability.Col.Component   = Pi\u00E8ce
 componentanalysisdlg.TabStability.Col.Mass        = Masse
-componentanalysisdlg.TabStability.ttip            = Information de stabilit\u00E9
-componentanalysisdlg.TitledBorder.warnings        = Avertissements:
-componentanalysisdlg.ToggleBut.worst              = Pire
+ComponentAnalysisGeneralTab.TabStability.ttip            = Information de stabilit\u00E9
+ComponentAnalysisGeneralTab.TitledBorder.warnings        = Avertissements:
+ComponentAnalysisGeneralTab.ToggleBut.worst              = Pire
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis            = Analyse pi\u00E8ce
-componentanalysisdlg.dragTabchar                  = Caract\u00E9ristiques de tra\u00EEn\u00E9e
-componentanalysisdlg.dragTabchar.ttip             = Caract\u00E9ristiques de tra\u00EEn\u00E9e
-componentanalysisdlg.dragTableModel.Col.Base      = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Component = Pi\u00E8ce
-componentanalysisdlg.dragTableModel.Col.Pressure  = <html>Pression C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction  = <html>Friction C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total     = <html>Total C<sub>D</sub>
-componentanalysisdlg.lbl.activestages             = Etages actifs:
-componentanalysisdlg.lbl.angleofattack            = Angle d'attaque:
-componentanalysisdlg.lbl.machnumber               = Mach number:
-componentanalysisdlg.lbl.motorconf                = Configuration moteur:
-componentanalysisdlg.lbl.refarea                  = Aire de r\u00E9f\u00E9rence: 
-componentanalysisdlg.lbl.reflenght                = Longueur de r\u00E9f\u00E9rence: 
-componentanalysisdlg.lbl.rollrate                 = Taux de roulis:
-componentanalysisdlg.lbl.winddir                  = Direction du vent:
-componentanalysisdlg.noWarnings                   = <html><i><font color="gray">Pas d'alertes.</font></i>
-componentanalysisdlg.println.closingmethod        = M\u00E9thode de cl\u00F4ture appel\u00E9e:
-componentanalysisdlg.println.settingnam           = SETTING NAN VALUES
-componentanalysisdlg.rollTableModel               = Dynamique du roulis
-componentanalysisdlg.rollTableModel.Col.component = Pi\u00E8ce
-componentanalysisdlg.rollTableModel.Col.rolldamp  = Coefficient amortisseur de roulis
-componentanalysisdlg.rollTableModel.Col.rollforc  = Coefficient de for\u00E7age de roulis
-componentanalysisdlg.rollTableModel.Col.total     = <html>Total C<sub>l</sub>
-componentanalysisdlg.rollTableModel.ttip          = Dynamique du roulis
+ComponentAnalysisDialog.componentanalysis            = Analyse pi\u00E8ce
+ComponentAnalysisGeneralTab.dragTabchar                  = Caract\u00E9ristiques de tra\u00EEn\u00E9e
+ComponentAnalysisGeneralTab.dragTabchar.ttip             = Caract\u00E9ristiques de tra\u00EEn\u00E9e
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base      = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Pi\u00E8ce
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure  = <html>Pression C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction  = <html>Friction C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total     = <html>Total C<sub>D</sub>
+ComponentAnalysisGeneralTab.lbl.activestages             = Etages actifs:
+ComponentAnalysisGeneralTab.lbl.angleofattack            = Angle d'attaque:
+ComponentAnalysisGeneralTab.lbl.machnumber               = Mach number:
+ComponentAnalysisGeneralTab.lbl.motorconf                = Configuration moteur:
+ComponentAnalysisGeneralTab.lbl.refarea                  = Aire de r\u00E9f\u00E9rence: 
+ComponentAnalysisGeneralTab.lbl.reflenght                = Longueur de r\u00E9f\u00E9rence: 
+ComponentAnalysisGeneralTab.lbl.rollrate                 = Taux de roulis:
+ComponentAnalysisGeneralTab.lbl.winddir                  = Direction du vent:
+ComponentAnalysisGeneralTab.noWarnings                   = <html><i><font color="gray">Pas d'alertes.</font></i>
+ComponentAnalysisGeneralTab.println.closingmethod        = M\u00E9thode de cl\u00F4ture appel\u00E9e:
+ComponentAnalysisGeneralTab.println.settingnam           = SETTING NAN VALUES
+ComponentAnalysisGeneralTab.rollTableModel               = Dynamique du roulis
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Pi\u00E8ce
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp  = Coefficient amortisseur de roulis
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc  = Coefficient de for\u00E7age de roulis
+ComponentAnalysisGeneralTab.rollTableModel.Col.total     = <html>Total C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel.ttip          = Dynamique du roulis
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir                  = Direction du vent
