+++ core/src/main/resources/l10n/messages_it.properties
@@ -567,37 +567,37 @@ compaddbuttons.askPosition.Addtotheend = Aggiungi alla fine
 compaddbuttons.askPosition.Cancel = Annulla
 
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis = Analisi del componente
-componentanalysisdlg.lbl.winddir = Direzione del vento:
-componentanalysisdlg.TitledBorder.warnings = Avvertimenti:
-componentanalysisdlg.ToggleBut.worst = Peggiore
-componentanalysisdlg.lbl.angleofattack = Angolo di attacco:
-componentanalysisdlg.lbl.machnumber = Numero di Mach:
-componentanalysisdlg.lbl.rollrate = Tasso di rollio:
-componentanalysisdlg.lbl.activestages = Stadi attivi:
-componentanalysisdlg.lbl.motorconf = Configurazione del motore:
-componentanalysisdlg.TabStability.Col = Componente
-componentanalysisdlg.TabStability = Stabilita'
-componentanalysisdlg.TabStability.ttip = Informazioni sulla stabilita'Stability information
-componentanalysisdlg.dragTableModel.Col.Component = Componente
-componentanalysisdlg.dragTableModel.Col.Pressure = <html>Pressione C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Base = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction = <html>Attrito C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total = <html>Totale C<sub>D</sub>
-componentanalysisdlg.dragTabchar = Caratteristiche di resistenza
-componentanalysisdlg.dragTabchar.ttip = Caratteristiche di resistenza
-componentanalysisdlg.rollTableModel.Col.component = Componente
-componentanalysisdlg.rollTableModel.Col.rollforc = Roll forcing coefficient
-componentanalysisdlg.rollTableModel.Col.rolldamp = Coefficiente smorzamento di rollio
-componentanalysisdlg.rollTableModel.Col.total = <html>Totale C<sub>l</sub>
-componentanalysisdlg.rollTableModel = Dinamiche di rollio
-componentanalysisdlg.rollTableModel.ttip = Dinamiche di rollio
-componentanalysisdlg.println.closingmethod = Sto chiudendo il metodo chiamato:
-componentanalysisdlg.println.settingnam = IMPOSTAZIONI DEI VALORI NaN 
-componentanalysisdlg.lbl.reflenght = Lunghezza di riferimento: 
-componentanalysisdlg.lbl.refarea = Area di riferimento: 
+ComponentAnalysisDialog.componentanalysis = Analisi del componente
+ComponentAnalysisGeneralTab.lbl.winddir = Direzione del vento:
+ComponentAnalysisGeneralTab.TitledBorder.warnings = Avvertimenti:
+ComponentAnalysisGeneralTab.ToggleBut.worst = Peggiore
+ComponentAnalysisGeneralTab.lbl.angleofattack = Angolo di attacco:
+ComponentAnalysisGeneralTab.lbl.machnumber = Numero di Mach:
+ComponentAnalysisGeneralTab.lbl.rollrate = Tasso di rollio:
+ComponentAnalysisGeneralTab.lbl.activestages = Stadi attivi:
+ComponentAnalysisGeneralTab.lbl.motorconf = Configurazione del motore:
+ComponentAnalysisGeneralTab.TabStability.Col = Componente
+ComponentAnalysisGeneralTab.TabStability = Stabilita'
+ComponentAnalysisGeneralTab.TabStability.ttip = Informazioni sulla stabilita'Stability information
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Componente
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure = <html>Pressione C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction = <html>Attrito C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total = <html>Totale C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTabchar = Caratteristiche di resistenza
+ComponentAnalysisGeneralTab.dragTabchar.ttip = Caratteristiche di resistenza
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Componente
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc = Roll forcing coefficient
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp = Coefficiente smorzamento di rollio
+ComponentAnalysisGeneralTab.rollTableModel.Col.total = <html>Totale C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel = Dinamiche di rollio
+ComponentAnalysisGeneralTab.rollTableModel.ttip = Dinamiche di rollio
+ComponentAnalysisGeneralTab.println.closingmethod = Sto chiudendo il metodo chiamato:
+ComponentAnalysisGeneralTab.println.settingnam = IMPOSTAZIONI DEI VALORI NaN 
+ComponentAnalysisGeneralTab.lbl.reflenght = Lunghezza di riferimento: 
+ComponentAnalysisGeneralTab.lbl.refarea = Area di riferimento: 
 !componentanalysisdlg.But.close =Chiudi
-componentanalysisdlg.TabStability.Col.Component = Componente
+ComponentAnalysisGeneralTab.TabStability.Col.Component = Componente
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir = Direzione del vento
