+++ core/src/main/resources/l10n/messages_ja.properties
@@ -592,42 +592,42 @@ compaddbuttons.askPosition.Addtotheend  = \u6700\u5F8C\u306B\u8FFD\u52A0
 compaddbuttons.askPosition.Cancel  = \u30AD\u30E3\u30F3\u30BB\u30EB
 
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis  = \u90E8\u54C1\u89E3\u6790
-componentanalysisdlg.lbl.winddir  = \u98A8\u5411\u304D\uFF1A
-componentanalysisdlg.TitledBorder.warnings  = \u30A8\u30E9\u30FC\uFF1A
-componentanalysisdlg.ToggleBut.worst  = Worst
-componentanalysisdlg.lbl.angleofattack  = \u8FCE\u3048\u89D2\uFF1A
-componentanalysisdlg.lbl.machnumber  = \u30DE\u30C3\u30CF\u6570\uFF1A
-componentanalysisdlg.lbl.rollrate  = \u89D2\u901F\u5EA6\uFF08\u30ED\u30FC\u30EB\uFF09\uFF1A
-componentanalysisdlg.lbl.activestages  = Active stages:
-componentanalysisdlg.lbl.motorconf  = Motor configuration:
-componentanalysisdlg.TabStability.Col  = \u90E8\u54C1
-componentanalysisdlg.TabStability.Col.CG  = CG
+ComponentAnalysisDialog.componentanalysis  = \u90E8\u54C1\u89E3\u6790
+ComponentAnalysisGeneralTab.lbl.winddir  = \u98A8\u5411\u304D\uFF1A
+ComponentAnalysisGeneralTab.TitledBorder.warnings  = \u30A8\u30E9\u30FC\uFF1A
+ComponentAnalysisGeneralTab.ToggleBut.worst  = Worst
+ComponentAnalysisGeneralTab.lbl.angleofattack  = \u8FCE\u3048\u89D2\uFF1A
+ComponentAnalysisGeneralTab.lbl.machnumber  = \u30DE\u30C3\u30CF\u6570\uFF1A
+ComponentAnalysisGeneralTab.lbl.rollrate  = \u89D2\u901F\u5EA6\uFF08\u30ED\u30FC\u30EB\uFF09\uFF1A
+ComponentAnalysisGeneralTab.lbl.activestages  = Active stages:
+ComponentAnalysisGeneralTab.lbl.motorconf  = Motor configuration:
+ComponentAnalysisGeneralTab.TabStability.Col  = \u90E8\u54C1
+ComponentAnalysisGeneralTab.TabStability.Col.CG  = CG
 componentanalysisdlg.TabStability.Col.Mass  = \u8CEA\u91CF
-componentanalysisdlg.TabStability.Col.CP  = CP
-componentanalysisdlg.TabStability  = \u5B89\u5B9A\u6027
-componentanalysisdlg.TabStability.ttip  = \u5B89\u5B9A\u6027
-componentanalysisdlg.dragTableModel.Col.Component  = \u90E8\u54C1
-componentanalysisdlg.dragTableModel.Col.Pressure  = <html>\u5727\u529B C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Base  = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction  = <html>\u6469\u64E6 C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total  = <html>Total C<sub>D</sub>
-componentanalysisdlg.dragTabchar  = \u6297\u529B\u6307\u6A19
-componentanalysisdlg.dragTabchar.ttip  = \u6297\u529B\u6307\u6A19
-componentanalysisdlg.rollTableModel.Col.component  = \u90E8\u54C1
-componentanalysisdlg.rollTableModel.Col.rollforc  = \u529B\u4FC2\u6570\uFF08\u30ED\u30FC\u30EB\uFF09
-componentanalysisdlg.rollTableModel.Col.rolldamp  = \u6E1B\u8870\u4FC2\u6570\uFF08\u30ED\u30FC\u30EB\uFF09
-componentanalysisdlg.rollTableModel.Col.total  = <html>Total C<sub>l</sub>
-componentanalysisdlg.rollTableModel  = Roll dynamics
-componentanalysisdlg.rollTableModel.ttip  = Roll dynamics
-componentanalysisdlg.println.closingmethod  = Closing method called:
-componentanalysisdlg.println.settingnam  = SETTING NAN VALUES
-componentanalysisdlg.lbl.reflenght  = \u4EE3\u8868\u9577\u3055\uFF1A
-componentanalysisdlg.lbl.refarea  = \u4EE3\u8868\u9762\u7A4D\uFF1A
+ComponentAnalysisGeneralTab.TabStability.Col.CP  = CP
+ComponentAnalysisGeneralTab.TabStability  = \u5B89\u5B9A\u6027
+ComponentAnalysisGeneralTab.TabStability.ttip  = \u5B89\u5B9A\u6027
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component  = \u90E8\u54C1
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure  = <html>\u5727\u529B C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base  = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction  = <html>\u6469\u64E6 C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total  = <html>Total C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTabchar  = \u6297\u529B\u6307\u6A19
+ComponentAnalysisGeneralTab.dragTabchar.ttip  = \u6297\u529B\u6307\u6A19
+ComponentAnalysisGeneralTab.rollTableModel.Col.component  = \u90E8\u54C1
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc  = \u529B\u4FC2\u6570\uFF08\u30ED\u30FC\u30EB\uFF09
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp  = \u6E1B\u8870\u4FC2\u6570\uFF08\u30ED\u30FC\u30EB\uFF09
+ComponentAnalysisGeneralTab.rollTableModel.Col.total  = <html>Total C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel  = Roll dynamics
+ComponentAnalysisGeneralTab.rollTableModel.ttip  = Roll dynamics
+ComponentAnalysisGeneralTab.println.closingmethod  = Closing method called:
+ComponentAnalysisGeneralTab.println.settingnam  = SETTING NAN VALUES
+ComponentAnalysisGeneralTab.lbl.reflenght  = \u4EE3\u8868\u9577\u3055\uFF1A
+ComponentAnalysisGeneralTab.lbl.refarea  = \u4EE3\u8868\u9762\u7A4D\uFF1A
 !componentanalysisdlg.But.close 
-componentanalysisdlg.TabStability.Col.Component  = \u90E8\u54C1
-componentanalysisdlg.TOTAL  = Total
-componentanalysisdlg.noWarnings  = <html><i><font color
+ComponentAnalysisGeneralTab.TabStability.Col.Component  = \u90E8\u54C1
+ComponentAnalysisGeneralTab.TOTAL  = Total
+ComponentAnalysisGeneralTab.noWarnings  = <html><i><font color
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir  = \u98A8\u5411\u304D\uFF1A
