+++ core/src/main/resources/l10n/messages_nl.properties
@@ -720,43 +720,43 @@ compaddbuttons.askPosition.Addtotheend = Aan het eind toevoegen
 compaddbuttons.askPosition.Cancel = Annuleer
 
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis = Onderdeelanalyses
-componentanalysisdlg.lbl.winddir = Windrichting:
-componentanalysisdlg.TitledBorder.warnings = Waarschuwingen:
-componentanalysisdlg.ToggleBut.worst = Ergste
-componentanalysisdlg.lbl.angleofattack = Invalshoek:
-componentanalysisdlg.lbl.machnumber = Machgetal:
-componentanalysisdlg.lbl.rollrate = Rolpercentage:
-componentanalysisdlg.lbl.activestages = Actieve trappen:
-componentanalysisdlg.lbl.motorconf = Motorconfiguratie:
-componentanalysisdlg.TabStability.Col = Onderdeelnaam
-componentanalysisdlg.TabStability.Col.EachMass = Massa Instantie
-componentanalysisdlg.TabStability.Col.AllMass = Massa Geheel
-componentanalysisdlg.TabStability.Col.CG = ZP Locatie Geheel
-componentanalysisdlg.TabStability.Col.CP = DP
-componentanalysisdlg.TabStability = Stabiliteit
-componentanalysisdlg.TabStability.ttip = Stabiliteitsinformatie
-componentanalysisdlg.dragTableModel.Col.Component = Onderdeel
-componentanalysisdlg.dragTableModel.Col.Pressure = <html>Druk C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Base = <html>Basis C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction = <html>Wrijving C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total = <html>Totaal C<sub>D</sub>
-componentanalysisdlg.dragTabchar = Weerstandskarakteristieken
-componentanalysisdlg.dragTabchar.ttip = Weerstandskarakteristieken
-componentanalysisdlg.rollTableModel.Col.component = Onderdeel
-componentanalysisdlg.rollTableModel.Col.rollforc = Rolverdringingscoëfficiënt
-componentanalysisdlg.rollTableModel.Col.rolldamp = Roldempingscoëfficiënt
-componentanalysisdlg.rollTableModel.Col.total = <html>Totaal C<sub>l</sub>
-componentanalysisdlg.rollTableModel = Roldynamieken
-componentanalysisdlg.rollTableModel.ttip = Roldynamieken
-componentanalysisdlg.println.closingmethod = Afsluitmethode genoemd:
-componentanalysisdlg.println.settingnam = INSTELLEN VAN NAN WAARDEN
-componentanalysisdlg.lbl.reflenght = Referentielengte:
-componentanalysisdlg.lbl.refarea = Referentie-oppervlakte:
+ComponentAnalysisDialog.componentanalysis = Onderdeelanalyses
+ComponentAnalysisGeneralTab.lbl.winddir = Windrichting:
+ComponentAnalysisGeneralTab.TitledBorder.warnings = Waarschuwingen:
+ComponentAnalysisGeneralTab.ToggleBut.worst = Ergste
+ComponentAnalysisGeneralTab.lbl.angleofattack = Invalshoek:
+ComponentAnalysisGeneralTab.lbl.machnumber = Machgetal:
+ComponentAnalysisGeneralTab.lbl.rollrate = Rolpercentage:
+ComponentAnalysisGeneralTab.lbl.activestages = Actieve trappen:
+ComponentAnalysisGeneralTab.lbl.motorconf = Motorconfiguratie:
+ComponentAnalysisGeneralTab.TabStability.Col = Onderdeelnaam
+ComponentAnalysisGeneralTab.TabStability.Col.EachMass = Massa Instantie
+ComponentAnalysisGeneralTab.TabStability.Col.AllMass = Massa Geheel
+ComponentAnalysisGeneralTab.TabStability.Col.CG = ZP Locatie Geheel
+ComponentAnalysisGeneralTab.TabStability.Col.CP = DP
+ComponentAnalysisGeneralTab.TabStability = Stabiliteit
+ComponentAnalysisGeneralTab.TabStability.ttip = Stabiliteitsinformatie
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Onderdeel
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure = <html>Druk C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base = <html>Basis C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction = <html>Wrijving C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total = <html>Totaal C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTabchar = Weerstandskarakteristieken
+ComponentAnalysisGeneralTab.dragTabchar.ttip = Weerstandskarakteristieken
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Onderdeel
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc = Rolverdringingscoëfficiënt
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp = Roldempingscoëfficiënt
+ComponentAnalysisGeneralTab.rollTableModel.Col.total = <html>Totaal C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel = Roldynamieken
+ComponentAnalysisGeneralTab.rollTableModel.ttip = Roldynamieken
+ComponentAnalysisGeneralTab.println.closingmethod = Afsluitmethode genoemd:
+ComponentAnalysisGeneralTab.println.settingnam = INSTELLEN VAN NAN WAARDEN
+ComponentAnalysisGeneralTab.lbl.reflenght = Referentielengte:
+ComponentAnalysisGeneralTab.lbl.refarea = Referentie-oppervlakte:
 !componentanalysisdlg.But.close = Sluit
-componentanalysisdlg.TabStability.Col.Component = Onderdeel
-componentanalysisdlg.TOTAL = Totaal
-componentanalysisdlg.noWarnings = <html><i><font color=\"gray\">Geen waarschuwingen.</font></i>
+ComponentAnalysisGeneralTab.TabStability.Col.Component = Onderdeel
+ComponentAnalysisGeneralTab.TOTAL = Totaal
+ComponentAnalysisGeneralTab.noWarnings = <html><i><font color=\"gray\">Geen waarschuwingen.</font></i>
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir = Windrichting
