+++ core/src/main/resources/l10n/messages_pl.properties
@@ -511,37 +511,37 @@ update.dlg.latestVersion = Korzystasz z najnowszej wersji OpenRocket: %s.
  compaddbuttons.askPosition.Cancel  =  Anuluj
  
  ! Component Analysis Dialog
- componentanalysisdlg.componentanalysis  =  Analiza cz\u0119\u015Bci
- componentanalysisdlg.lbl.winddir  =  Kierunek wiatru:
- componentanalysisdlg.TitledBorder.warnings  =  Ostrze\u017Cenia:
- componentanalysisdlg.ToggleBut.worst  =  Najgorszy
- componentanalysisdlg.lbl.angleofattack  =  K\u0105t natarcia:
- componentanalysisdlg.lbl.machnumber  =  Liczba Macha:
- componentanalysisdlg.lbl.rollrate  =  Tempo obrotu:
- componentanalysisdlg.lbl.activestages  =  Aktywne stopnie:
- componentanalysisdlg.lbl.motorconf  =  Konfiguracja silnika:
- componentanalysisdlg.TabStability.Col  =  Cz\u0119\u015B\u0107
- componentanalysisdlg.TabStability  =  Stabilno\u015B\u0107
- componentanalysisdlg.TabStability.ttip  =  Informacje dotycz\u0105ce stabilno\u015Bci
- componentanalysisdlg.dragTableModel.Col.Component  =  Cz\u0119\u015B\u0107
- componentanalysisdlg.dragTableModel.Col.Pressure  =  <html>Ci\u015Bnienie C<sub>D</sub>
- componentanalysisdlg.dragTableModel.Col.Base  =  <html>Baza C<sub>D</sub>
- componentanalysisdlg.dragTableModel.Col.friction  =  <html>Tarcie C<sub>D</sub>
- componentanalysisdlg.dragTableModel.Col.total  =  <html>Razem C<sub>D</sub>
- componentanalysisdlg.dragTabchar  =  Charakterystyka oporu
- componentanalysisdlg.dragTabchar.ttip  =  Charakterystyka oporu
- componentanalysisdlg.rollTableModel.Col.component  =  Cz\u0119\u015B\u0107
- componentanalysisdlg.rollTableModel.Col.rollforc  =  Wspó\u0142czynnik wymuszania obrotu
- componentanalysisdlg.rollTableModel.Col.rolldamp  =  Wspó\u0142czynnik hamowania obrotu
- componentanalysisdlg.rollTableModel.Col.total  =  <html>Razem C<sub>I</sub>
- componentanalysisdlg.rollTableModel  =  Dynamika obrotu
- componentanalysisdlg.rollTableModel.ttip  =  Dynamika obrotu
- componentanalysisdlg.println.closingmethod  =  Metoda zamykania:
- componentanalysisdlg.println.settingnam  =  USTAWIANIE WARTO\u015ACI NAN
- componentanalysisdlg.lbl.reflenght  =  D\u0142ugo\u015B\u0107 wzorcowa:
- componentanalysisdlg.lbl.refarea  =  Powierzchnia wzorcowa:
+ ComponentAnalysisDialog.componentanalysis  =  Analiza cz\u0119\u015Bci
+ ComponentAnalysisGeneralTab.lbl.winddir  =  Kierunek wiatru:
+ ComponentAnalysisGeneralTab.TitledBorder.warnings  =  Ostrze\u017Cenia:
+ ComponentAnalysisGeneralTab.ToggleBut.worst  =  Najgorszy
+ ComponentAnalysisGeneralTab.lbl.angleofattack  =  K\u0105t natarcia:
+ ComponentAnalysisGeneralTab.lbl.machnumber  =  Liczba Macha:
+ ComponentAnalysisGeneralTab.lbl.rollrate  =  Tempo obrotu:
+ ComponentAnalysisGeneralTab.lbl.activestages  =  Aktywne stopnie:
+ ComponentAnalysisGeneralTab.lbl.motorconf  =  Konfiguracja silnika:
+ ComponentAnalysisGeneralTab.TabStability.Col  =  Cz\u0119\u015B\u0107
+ ComponentAnalysisGeneralTab.TabStability  =  Stabilno\u015B\u0107
+ ComponentAnalysisGeneralTab.TabStability.ttip  =  Informacje dotycz\u0105ce stabilno\u015Bci
+ ComponentAnalysisGeneralTab.dragTableModel.Col.Component  =  Cz\u0119\u015B\u0107
+ ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure  =  <html>Ci\u015Bnienie C<sub>D</sub>
+ ComponentAnalysisGeneralTab.dragTableModel.Col.Base  =  <html>Baza C<sub>D</sub>
+ ComponentAnalysisGeneralTab.dragTableModel.Col.friction  =  <html>Tarcie C<sub>D</sub>
+ ComponentAnalysisGeneralTab.dragTableModel.Col.total  =  <html>Razem C<sub>D</sub>
+ ComponentAnalysisGeneralTab.dragTabchar  =  Charakterystyka oporu
+ ComponentAnalysisGeneralTab.dragTabchar.ttip  =  Charakterystyka oporu
+ ComponentAnalysisGeneralTab.rollTableModel.Col.component  =  Cz\u0119\u015B\u0107
+ ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc  =  Wspó\u0142czynnik wymuszania obrotu
+ ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp  =  Wspó\u0142czynnik hamowania obrotu
+ ComponentAnalysisGeneralTab.rollTableModel.Col.total  =  <html>Razem C<sub>I</sub>
+ ComponentAnalysisGeneralTab.rollTableModel  =  Dynamika obrotu
+ ComponentAnalysisGeneralTab.rollTableModel.ttip  =  Dynamika obrotu
+ ComponentAnalysisGeneralTab.println.closingmethod  =  Metoda zamykania:
+ ComponentAnalysisGeneralTab.println.settingnam  =  USTAWIANIE WARTO\u015ACI NAN
+ ComponentAnalysisGeneralTab.lbl.reflenght  =  D\u0142ugo\u015B\u0107 wzorcowa:
+ ComponentAnalysisGeneralTab.lbl.refarea  =  Powierzchnia wzorcowa:
  !componentanalysisdlg.But.close 
- componentanalysisdlg.TabStability.Col.Component  =  Cz\u0119\u015B\u0107
+ ComponentAnalysisGeneralTab.TabStability.Col.Component  =  Cz\u0119\u015B\u0107
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir  =  Kierunek wiatru
