+++ core/src/main/resources/l10n/messages_pt.properties
@@ -1239,43 +1239,43 @@ compaddbuttons.askPosition.Inserthere  = Inserir aqui
 compaddbuttons.lbl.Youcanchange        = Voc\u00ea pode alterar a opera\u00e7\u00e3o padr\u00e3o em Prefer\u00eancias.
 compaddbuttons.lbl.insertcomp          = Inserir o componente depois do componente de corrente ou como o \u00faltimo componente?
 
-componentanalysisdlg.TOTAL                        = Total
-componentanalysisdlg.TabStability                 = Estabilidade
-componentanalysisdlg.TabStability.Col             = Componente
-componentanalysisdlg.TabStability.Col.CG          = CG
-componentanalysisdlg.TabStability.Col.CP          = CP
+ComponentAnalysisGeneralTab.TOTAL                        = Total
+ComponentAnalysisGeneralTab.TabStability                 = Estabilidade
+ComponentAnalysisGeneralTab.TabStability.Col             = Componente
+ComponentAnalysisGeneralTab.TabStability.Col.CG          = CG
+ComponentAnalysisGeneralTab.TabStability.Col.CP          = CP
 # componentanalysisdlg.But.close =Close
-componentanalysisdlg.TabStability.Col.Component   = Componente
+ComponentAnalysisGeneralTab.TabStability.Col.Component   = Componente
 componentanalysisdlg.TabStability.Col.Mass        = Massa.
-componentanalysisdlg.TabStability.ttip            = Informa\u00e7\u00f5es sobre a estabilidade
-componentanalysisdlg.TitledBorder.warnings        = Avisos:
-componentanalysisdlg.ToggleBut.worst              = Pior
+ComponentAnalysisGeneralTab.TabStability.ttip            = Informa\u00e7\u00f5es sobre a estabilidade
+ComponentAnalysisGeneralTab.TitledBorder.warnings        = Avisos:
+ComponentAnalysisGeneralTab.ToggleBut.worst              = Pior
 # Component Analysis Dialog
-componentanalysisdlg.componentanalysis            = An\u00e1lise dos componentes
-componentanalysisdlg.dragTabchar                  = Caracter\u00edsticas do arrasto
-componentanalysisdlg.dragTabchar.ttip             = Caracter\u00edsticas do arrasto
-componentanalysisdlg.dragTableModel.Col.Base      = <html>Base C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Component = Componente
-componentanalysisdlg.dragTableModel.Col.Pressure  = <html>Press\u00e3o C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction  = <html>Atrito C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total     = <html>Total C<sub>D</sub>
-componentanalysisdlg.lbl.activestages             = Est\u00e1gios ativos:
-componentanalysisdlg.lbl.angleofattack            = \u00c2ngulo de ataque:
-componentanalysisdlg.lbl.machnumber               = N\u00famero de Mach:
-componentanalysisdlg.lbl.motorconf                = Configura\u00e7\u00e3o do motor:
-componentanalysisdlg.lbl.refarea                  = \u00c1rea de refer\u00eancia:
-componentanalysisdlg.lbl.reflenght                = Comprimento de refer\u00eancia:
-componentanalysisdlg.lbl.rollrate                 = Taxa de rota\u00e7\u00e3o:
-componentanalysisdlg.lbl.winddir                  = Dire\u00e7\u00e3o do vento:
-componentanalysisdlg.noWarnings                   = <html><i><font color="gray">Nenhuma advert\u00eancia.</font></i>
-componentanalysisdlg.println.closingmethod        = Fechando m\u00e9todo chamado:
-componentanalysisdlg.println.settingnam           = DEFININDO VALORES DE NAN
-componentanalysisdlg.rollTableModel               = Rota\u00e7\u00e3o din\u00e2mica
-componentanalysisdlg.rollTableModel.Col.component = Componente
-componentanalysisdlg.rollTableModel.Col.rolldamp  = Coeficiente de amortecimento de rota\u00e7\u00e3o
-componentanalysisdlg.rollTableModel.Col.rollforc  = Coeficiente for\u00e7ado de rota\u00e7\u00e3o
-componentanalysisdlg.rollTableModel.Col.total     = <html>Total C<sub>l</sub>
-componentanalysisdlg.rollTableModel.ttip          = Rota\u00e7\u00e3o din\u00e2mica
+ComponentAnalysisDialog.componentanalysis            = An\u00e1lise dos componentes
+ComponentAnalysisGeneralTab.dragTabchar                  = Caracter\u00edsticas do arrasto
+ComponentAnalysisGeneralTab.dragTabchar.ttip             = Caracter\u00edsticas do arrasto
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base      = <html>Base C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = Componente
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure  = <html>Press\u00e3o C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction  = <html>Atrito C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total     = <html>Total C<sub>D</sub>
+ComponentAnalysisGeneralTab.lbl.activestages             = Est\u00e1gios ativos:
+ComponentAnalysisGeneralTab.lbl.angleofattack            = \u00c2ngulo de ataque:
+ComponentAnalysisGeneralTab.lbl.machnumber               = N\u00famero de Mach:
+ComponentAnalysisGeneralTab.lbl.motorconf                = Configura\u00e7\u00e3o do motor:
+ComponentAnalysisGeneralTab.lbl.refarea                  = \u00c1rea de refer\u00eancia:
+ComponentAnalysisGeneralTab.lbl.reflenght                = Comprimento de refer\u00eancia:
+ComponentAnalysisGeneralTab.lbl.rollrate                 = Taxa de rota\u00e7\u00e3o:
+ComponentAnalysisGeneralTab.lbl.winddir                  = Dire\u00e7\u00e3o do vento:
+ComponentAnalysisGeneralTab.noWarnings                   = <html><i><font color="gray">Nenhuma advert\u00eancia.</font></i>
+ComponentAnalysisGeneralTab.println.closingmethod        = Fechando m\u00e9todo chamado:
+ComponentAnalysisGeneralTab.println.settingnam           = DEFININDO VALORES DE NAN
+ComponentAnalysisGeneralTab.rollTableModel               = Rota\u00e7\u00e3o din\u00e2mica
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = Componente
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp  = Coeficiente de amortecimento de rota\u00e7\u00e3o
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc  = Coeficiente for\u00e7ado de rota\u00e7\u00e3o
+ComponentAnalysisGeneralTab.rollTableModel.Col.total     = <html>Total C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel.ttip          = Rota\u00e7\u00e3o din\u00e2mica
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir                  = Dire\u00e7\u00e3o do vento
