+++ core/src/main/resources/l10n/messages_zh_CN.properties
@@ -1359,43 +1359,43 @@ compaddbuttons.askPosition.Inserthere  = \u5728\u6B64\u63D2\u5165
 compaddbuttons.lbl.Youcanchange        = \u4F60\u53EF\u4EE5\u5728\u9996\u9009\u9879\u4E2D\u8BBE\u5B9A\u9ED8\u8BA4\u64CD\u4F5C.
 compaddbuttons.lbl.insertcomp          = \u7D27\u63A5\u5F53\u524D\u7EC4\u4EF6\u6DFB\u52A0\u65B0\u7EC4\u4EF6\u6216\u5728\u5C3E\u90E8\u6DFB\u52A0\u65B0\u7EC4\u4EF6
 
-componentanalysisdlg.TOTAL                        = \u603B\u5171
-componentanalysisdlg.TabStability                 = \u7A33\u5B9A\u6027
-componentanalysisdlg.TabStability.Col             = \u7EC4\u4EF6
-componentanalysisdlg.TabStability.Col.CG          = \u91CD\u5FC3
-componentanalysisdlg.TabStability.Col.CP          = \u538B\u5FC3
+ComponentAnalysisGeneralTab.TOTAL                        = \u603B\u5171
+ComponentAnalysisGeneralTab.TabStability                 = \u7A33\u5B9A\u6027
+ComponentAnalysisGeneralTab.TabStability.Col             = \u7EC4\u4EF6
+ComponentAnalysisGeneralTab.TabStability.Col.CG          = \u91CD\u5FC3
+ComponentAnalysisGeneralTab.TabStability.Col.CP          = \u538B\u5FC3
 !componentanalysisdlg.But.close  = \u5173\u95ed
-componentanalysisdlg.TabStability.Col.Component   = \u7EC4\u4EF6
+ComponentAnalysisGeneralTab.TabStability.Col.Component   = \u7EC4\u4EF6
 componentanalysisdlg.TabStability.Col.Mass        = \u8D28\u91CF
-componentanalysisdlg.TabStability.ttip            = \u7A33\u5B9A\u6027\u4FE1\u606F
-componentanalysisdlg.TitledBorder.warnings        = \u8B66\u544A:
-componentanalysisdlg.ToggleBut.worst              = \u6700\u574F\u60C5\u51B5
+ComponentAnalysisGeneralTab.TabStability.ttip            = \u7A33\u5B9A\u6027\u4FE1\u606F
+ComponentAnalysisGeneralTab.TitledBorder.warnings        = \u8B66\u544A:
+ComponentAnalysisGeneralTab.ToggleBut.worst              = \u6700\u574F\u60C5\u51B5
 ! Component Analysis Dialog
-componentanalysisdlg.componentanalysis            = \u7EC4\u4EF6\u5206\u6790
-componentanalysisdlg.dragTabchar                  = \u963B\u529B\u7279\u6027
-componentanalysisdlg.dragTabchar.ttip             = \u963B\u529B\u7279\u6027
-componentanalysisdlg.dragTableModel.Col.Base      = <html>\u57FA\u51C6 C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.Component = \u7EC4\u4EF6
-componentanalysisdlg.dragTableModel.Col.Pressure  = <html>\u538B\u529B C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.friction  = <html>\u6469\u64E6 C<sub>D</sub>
-componentanalysisdlg.dragTableModel.Col.total     = <html>\u603B C<sub>D</sub>
-componentanalysisdlg.lbl.activestages             = \u6709\u6548\u7EA7:
-componentanalysisdlg.lbl.angleofattack            = \u653B\u89D2:
-componentanalysisdlg.lbl.machnumber               = \u9A6C\u8D6B:
-componentanalysisdlg.lbl.motorconf                = \u53D1\u52A8\u673A\u8BBE\u7F6E:
-componentanalysisdlg.lbl.refarea                  = \u53C2\u8003\u9762\u79EF:
-componentanalysisdlg.lbl.reflenght                = \u53C2\u8003\u957F\u5EA6:
-componentanalysisdlg.lbl.rollrate                 = \u89D2\u901F\u5EA6:
-componentanalysisdlg.lbl.winddir                  = \u98CE\u5411:
-componentanalysisdlg.noWarnings                   = <html><i><font color="gray">\u6CA1\u6709\u8B66\u544A.</font></i>
-componentanalysisdlg.println.closingmethod        = \u8C03\u7528\u5173\u95ED\u4E8B\u4EF6:
-componentanalysisdlg.println.settingnam           = \u8BBE\u7F6E\u975E\u6570\u5B57\u503C
-componentanalysisdlg.rollTableModel               = \u6EDA\u8F6C\u52A8\u529B\u5B66
-componentanalysisdlg.rollTableModel.Col.component = \u7EC4\u4EF6
-componentanalysisdlg.rollTableModel.Col.rolldamp  = \u6EDA\u8F6C\u963B\u5C3C\u7CFB\u6570
-componentanalysisdlg.rollTableModel.Col.rollforc  = \u6EDA\u8F6C\u529B\u7CFB\u6570
-componentanalysisdlg.rollTableModel.Col.total     = <html>\u603B C<sub>l</sub>
-componentanalysisdlg.rollTableModel.ttip          = \u6EDA\u8F6C\u52A8\u529B\u5B66
+ComponentAnalysisDialog.componentanalysis            = \u7EC4\u4EF6\u5206\u6790
+ComponentAnalysisGeneralTab.dragTabchar                  = \u963B\u529B\u7279\u6027
+ComponentAnalysisGeneralTab.dragTabchar.ttip             = \u963B\u529B\u7279\u6027
+ComponentAnalysisGeneralTab.dragTableModel.Col.Base      = <html>\u57FA\u51C6 C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.Component = \u7EC4\u4EF6
+ComponentAnalysisGeneralTab.dragTableModel.Col.Pressure  = <html>\u538B\u529B C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.friction  = <html>\u6469\u64E6 C<sub>D</sub>
+ComponentAnalysisGeneralTab.dragTableModel.Col.total     = <html>\u603B C<sub>D</sub>
+ComponentAnalysisGeneralTab.lbl.activestages             = \u6709\u6548\u7EA7:
+ComponentAnalysisGeneralTab.lbl.angleofattack            = \u653B\u89D2:
+ComponentAnalysisGeneralTab.lbl.machnumber               = \u9A6C\u8D6B:
+ComponentAnalysisGeneralTab.lbl.motorconf                = \u53D1\u52A8\u673A\u8BBE\u7F6E:
+ComponentAnalysisGeneralTab.lbl.refarea                  = \u53C2\u8003\u9762\u79EF:
+ComponentAnalysisGeneralTab.lbl.reflenght                = \u53C2\u8003\u957F\u5EA6:
+ComponentAnalysisGeneralTab.lbl.rollrate                 = \u89D2\u901F\u5EA6:
+ComponentAnalysisGeneralTab.lbl.winddir                  = \u98CE\u5411:
+ComponentAnalysisGeneralTab.noWarnings                   = <html><i><font color="gray">\u6CA1\u6709\u8B66\u544A.</font></i>
+ComponentAnalysisGeneralTab.println.closingmethod        = \u8C03\u7528\u5173\u95ED\u4E8B\u4EF6:
+ComponentAnalysisGeneralTab.println.settingnam           = \u8BBE\u7F6E\u975E\u6570\u5B57\u503C
+ComponentAnalysisGeneralTab.rollTableModel               = \u6EDA\u8F6C\u52A8\u529B\u5B66
+ComponentAnalysisGeneralTab.rollTableModel.Col.component = \u7EC4\u4EF6
+ComponentAnalysisGeneralTab.rollTableModel.Col.rolldamp  = \u6EDA\u8F6C\u963B\u5C3C\u7CFB\u6570
+ComponentAnalysisGeneralTab.rollTableModel.Col.rollforc  = \u6EDA\u8F6C\u529B\u7CFB\u6570
+ComponentAnalysisGeneralTab.rollTableModel.Col.total     = <html>\u603B C<sub>l</sub>
+ComponentAnalysisGeneralTab.rollTableModel.ttip          = \u6EDA\u8F6C\u52A8\u529B\u5B66
 
 ! CADomainDataType
 CADomainDataType.lbl.winddir                  = \u98CE\u5411
